backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.5
pydantic-settings==2.8.1
PyJWT==2.9.0
python-multipart==0.0.20
pdfplumber==0.11.5
pypdf==5.3.1
python-docx==1.1.2
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
